perf: dont fetch staking opportunities on app boot
